requirements.txt
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-docx>=1.1.0
openpyxl>=3.1.0
tqdm>=4.66.0
